+++ java/access_rules.go
@@ -23,6 +23,9 @@ const (
 // Default Timeout value for Create
 const WaitForAccessRuleTimeout = time.Duration(10 * time.Second)
 
+// Default Poll Interval value for Create
+const WaitForAccessRulePollInterval = time.Duration(1 * time.Second)
+
 // AccessRules returns a UtilityClient for managing SSH Keys and Access Rules for a JaaS Service Instance
 func (c *JavaClient) AccessRules() *UtilityClient {
 	return &UtilityClient{
@@ -132,6 +135,8 @@ type CreateAccessRuleInput struct {
 	// Desired Status of the rule. Either "disabled" or "enabled".
 	// Required
 	Status AccessRuleStatus `json:"status"`
+	// Time to wait between polling for access rule to be ready
+	PollInterval time.Duration `json:"-"`
 	// Time to wait for an access rule to be ready
 	Timeout time.Duration `json:"-"`
 }
@@ -150,6 +155,11 @@ func (c *UtilityClient) CreateAccessRule(input *CreateAccessRuleInput) (*AccessR
 		return nil, err
 	}
 
+	pollInterval := input.PollInterval
+	if pollInterval == 0 {
+		pollInterval = WaitForAccessRulePollInterval
+	}
+
 	timeout := input.Timeout
 	if timeout == 0 {
 		timeout = WaitForAccessRuleTimeout
@@ -159,7 +169,7 @@ func (c *UtilityClient) CreateAccessRule(input *CreateAccessRuleInput) (*AccessR
 		Name: input.Name,
 	}
 
-	result, err := c.WaitForAccessRuleReady(getInput, timeout)
+	result, err := c.WaitForAccessRuleReady(getInput, pollInterval, timeout)
 	if err != nil {
 		return nil, err
 	}
@@ -267,6 +277,8 @@ type DeleteAccessRuleInput struct {
 	// modified on an access rule.
 	// Required
 	Status AccessRuleStatus `json:"status"`
+	// Time to wait between polling for access rule to be ready
+	PollInterval time.Duration `json:"-"`
 	// Time to wait for an access rule to be ready
 	Timeout time.Duration `json:"-"`
 }
@@ -288,6 +300,11 @@ func (c *UtilityClient) DeleteAccessRule(input *DeleteAccessRuleInput) error {
 		return err
 	}
 
+	pollInterval := input.PollInterval
+	if pollInterval == 0 {
+		pollInterval = WaitForAccessRulePollInterval
+	}
+
 	timeout := input.Timeout
 	if timeout == 0 {
 		timeout = WaitForAccessRuleTimeout
@@ -297,7 +314,7 @@ func (c *UtilityClient) DeleteAccessRule(input *DeleteAccessRuleInput) error {
 		Name: input.Name,
 	}
 
-	_, err := c.WaitForAccessRuleDeleted(getInput, timeout)
+	_, err := c.WaitForAccessRuleDeleted(getInput, pollInterval, timeout)
 	if err != nil {
 		return err
 	}
@@ -305,10 +322,10 @@ func (c *UtilityClient) DeleteAccessRule(input *DeleteAccessRuleInput) error {
 	return nil
 }
 
-func (c *UtilityClient) WaitForAccessRuleReady(input *GetAccessRuleInput, timeout time.Duration) (*AccessRuleInfo, error) {
+func (c *UtilityClient) WaitForAccessRuleReady(input *GetAccessRuleInput, pollInterval, timeout time.Duration) (*AccessRuleInfo, error) {
 	var info *AccessRuleInfo
 	var getErr error
-	err := c.client.WaitFor("access rule to be ready", timeout, func() (bool, error) {
+	err := c.client.WaitFor("access rule to be ready", pollInterval, timeout, func() (bool, error) {
 		info, getErr = c.GetAccessRule(input)
 		if getErr != nil {
 			return false, getErr
@@ -323,10 +340,10 @@ func (c *UtilityClient) WaitForAccessRuleReady(input *GetAccessRuleInput, timeou
 	return info, err
 }
 
-func (c *UtilityClient) WaitForAccessRuleDeleted(input *GetAccessRuleInput, timeout time.Duration) (*AccessRuleInfo, error) {
+func (c *UtilityClient) WaitForAccessRuleDeleted(input *GetAccessRuleInput, pollInterval, timeout time.Duration) (*AccessRuleInfo, error) {
 	var info *AccessRuleInfo
 	var getErr error
-	err := c.client.WaitFor("access rule to be deleted", timeout, func() (bool, error) {
+	err := c.client.WaitFor("access rule to be deleted", pollInterval, timeout, func() (bool, error) {
 		info, getErr = c.GetAccessRule(input)
 		if getErr != nil {
 			return true, nil