+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## 0.7.1 (Unreleased)
 
+* java: New feature Java Access Rules [GH-112]
+
+* application: New feature Application Container [GH-113]
+
 ## 0.7.0 (February 15, 2018)
 
-* java: Migrated java service instance to version 1.1 of their api
+* java: Migrated java service instance to version 1.1 of their api [GH-107]
 
 ## 0.6.9 (February 14, 2018)